+++ agentops/__init__.py
@@ -1,4 +1,5 @@
-from .agentops import Client
-from .event import Event, EventState, SessionState
+from .client import Client
+from .event import Event, EventState
+from .session import SessionState
 from .config import Configuration
 from .logger import AgentOpsLogger
